requirements.txt
Django~=5.0.3
psycopg2-binary==2.9.9
pillow==10.2.0
django-tailwind[reload]
openpyxl
djangorestframework